+++ decorator_include.py
@@ -9,8 +9,7 @@ from os import path
 
 import pkg_resources
 
-from django.core.exceptions import ImproperlyConfigured
-from django.urls import URLPattern, URLResolver
+from django.urls import URLPattern, URLResolver, include
 from django.utils.functional import cached_property
 
 
@@ -82,6 +81,10 @@ class DecoratedPatterns(object):
         else:
             return self.urlconf
 
+    @cached_property
+    def app_name(self):
+        return getattr(self.urlconf_module, 'app_name', None)
+
 
 def decorator_include(decorators, arg, namespace=None):
     """
@@ -89,22 +92,12 @@ def decorator_include(decorators, arg, namespace=None):
     or an iterable of view decorators as the first argument and applies them,
     in reverse order, to all views in the included urlconf.
     """
-    app_name = None
-    if isinstance(arg, tuple):
-        # callable returning a namespace hint
-        try:
-            urlconf, app_name = arg
-        except ValueError:
-            if namespace:
-                raise ImproperlyConfigured(
-                    'Cannot override the namespace for a dynamic module that provides a namespace'
-                )
-            # can happen for example when using decorator_include with ``admin.site.urls``
-            urlconf, app_name, namespace = arg
+    if isinstance(arg, tuple) and len(arg) == 3 and not isinstance(arg[0], str):
+        # Special case where the function is used for something like `admin.site.urls`, which
+        # returns a tuple with the object containing the urls, the app name, and the namespace
+        # `include` does not support this pattern (you pass directly `admin.site.urls`, without
+        # using `include`) but we have to
+        urlconf_module, app_name, namespace = arg
     else:
-        # No namespace hint - use manually provided namespace
-        urlconf = arg
-
-    decorated_urlconf = DecoratedPatterns(urlconf, decorators)
-    namespace = namespace or app_name
-    return (decorated_urlconf, app_name, namespace)
+        urlconf_module, app_name, namespace = include(arg, namespace=namespace)
+    return DecoratedPatterns(urlconf_module, decorators), app_name, namespace