+++ CHANGELOG.rst
@@ -1,8 +1,9 @@
 Changelog
 =========
 
-UNRELEASED
-----------
+UNRELEASED *v3.0*
+-----------------
+* Handle namespace/app_name the same way as Django `include` function
 * Added support for Django 2.2.
 
 Release *v2.1* - ``2018-11-23``